Big change: Switch internal precision to double
git-svn-id: svn+ssh://erebor.esa.informatik.tu-darmstadt.de/home/wimi/svn/GaalopCompiler/gaalop-2.0/branches/tbaAndGapp@1069 4235ee25-d3ef-403b-b6ed-692c26853d4b

+++ api/src/main/java/de/gaalop/gapp/ConstantSetVectorArgument.java
@@ -6,9 +6,9 @@ package de.gaalop.gapp;
  */
 public class ConstantSetVectorArgument extends SetVectorArgument {
 
-    private float value;
+    private double value;
 
-    public ConstantSetVectorArgument(float value) {
+    public ConstantSetVectorArgument(double value) {
         this.value = value;
     }
 
@@ -17,11 +17,11 @@ public class ConstantSetVectorArgument extends SetVectorArgument {
         return true;
     }
 
-    public float getValue() {
+    public double getValue() {
         return value;
     }
 
-    public void setValue(float value) {
+    public void setValue(double value) {
         this.value = value;
     }
