+++ codegen-compressed/src/main/java/de/gaalop/compressed/BladePrinter.java
@@ -89,7 +89,7 @@ public class BladePrinter implements ExpressionVisitor {
 
 	@Override
 	public void visit(FloatConstant floatConstant) {
-		code.append(Float.toString(floatConstant.getValue()).replace('E', 'e'));
+		code.append(Double.toString(floatConstant.getValue()).replace('E', 'e'));
 	}
 
 	@Override
